Refactored a lot of the render testing code. Seperated overlays and layers. Overlays are now for UI and layers are for gameplay and Rendering
Fixed a lot of warnings as well.

+++ Engine/Classes/Input/InputHandler.h
@@ -12,16 +12,16 @@ inline namespace MARS
 		inline static bool IsKeyPressed(int32 Key) { return Instance->IsKeyPressed_Implementation(Key); }
 		inline static bool IsMouseButtonPressed(int32 Button) { return Instance->IsMouseButtonPressed_Implementation(Button); }
 		inline static std::pair<double, double> GetMousePos() { return Instance->GetMousePos_Implementation(); }
-		inline static float GetMouseX() { return Instance->GetMouseX_Implementation(); }
-		inline static float GetMouseY() { return Instance->GetMouseX_Implementation(); }
+		inline static double GetMouseX() { return Instance->GetMouseX_Implementation(); }
+		inline static double GetMouseY() { return Instance->GetMouseX_Implementation(); }
 
 	protected:
 
 		virtual bool IsMouseButtonPressed_Implementation(int32 Button) = 0;
 		virtual bool IsKeyPressed_Implementation(int32 Key) = 0;
 		virtual std::pair<double, double> GetMousePos_Implementation() = 0;
-		virtual float GetMouseX_Implementation() = 0;
-		virtual float GetMouseY_Implementation() = 0;
+		virtual double GetMouseX_Implementation() = 0;
+		virtual double GetMouseY_Implementation() = 0;
 
 	private:
