+++ Engine/Platform/Windows/Windows_Input.h
@@ -13,7 +13,7 @@ inline namespace MARS
 		virtual bool IsKeyPressed_Implementation(int32 Key) override;
 		virtual bool IsMouseButtonPressed_Implementation(int32 Button) override;
 		virtual std::pair<double, double> GetMousePos_Implementation() override;
-		virtual float GetMouseX_Implementation() override;
-		virtual float GetMouseY_Implementation() override;
+		virtual double GetMouseX_Implementation() override;
+		virtual double GetMouseY_Implementation() override;
 	};
 }
